Update script 2-largest_number.c

+++ 0x03-debugging/main.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-int largest_number(int, int, int);
+double largest_number(double, double, double);
 
 #endif /* MAIN_H */
